open .pck file with x,y and match to select data
the .pck must be a dictionary with an array of x motor values and an
array of y motoro values, and then options provided for selecting the
indeces of these arrays correpsonding to matches with circle or square
regions around the selected platemap samples. negative sample_no used
for background.
MOSTLY UNTESTED